Merge with EOM 1.44
author: angelos
Just to be on the safe side, use a struct stat.

author: angelos
Only do the secrecy check and parse the configuration file if it
actually exists.

author: angelos
Actually create all the pre-configured Transforms and Suites, even if
the user doesn't actually define them in the configuration file; ugly
kludge, but it allows use of isakmpd without a configuration file.

author: angelos
Add RIPEMD negotiation/configuration.