Factorize most of the pmap_bootstrap() guts used by pmap_motorola users, with a few hooks to cope with each architecture's specifics.
The new arch/m68k/m68k/pmap_bootstrap.c is not a standalone file, but will
be #included by the existing pmap_bootstrap.c code.

Tested on hp300 and mvme68k, mac68k coming soon. amiga will be left out
for now because it is a bit too different.