On some architectures, being too optimistic in preserve_subexpressions_p(), for example when -fexpensive-optimizations is used, sometimes causes incorrect code sequences to be generated.
As -fexpensive-optimizations is implied by -O2, this means we are potentially
affected by this in numerous places... so add a specific workaround to nullify
expensizve-optimizations in preserve_subexpressions_p() on a
known-to-be-affected basis.

So far, only alpha, powerpc, m88k, sparc and sparc64 are affected.

Problem tracked down by henning@ - analysis by etoh@ and I - ugly workaround
by me - ok deraadt@