Disable optimize_reg_copy_3() on m68k platforms.
This is a workaround for lines 1055-1057 of tcp_input.c being miscompiled,
though the problem looks like missing use/clobber qualifiers in the
extendplussidi insn, rather than a specific optimize_reg_copy_3() issue.

A proper fix may be devised in the future, in the meantime this allows
m68k platforms to be back in track.