another go at cleaning up bus_dma usage.
when you're using the bus dma api like we are, there are two
bus_dma_segment_t's involved, one for the pa stuff and one for the dva
stuff. we set this up ok in ami_allocmem, but then we forget half of it
when we return from that function. even worse, we assume theyre the same
thing in ami_freemem.

this diff fixes that by wrapping all the dma bits into a struct (called
struct ami_mem) and providing a few macros for accessing the useful parts
of it. we just pass this struct around now instead of having to handle all
the busdma bits in every function.

ok marco@ jason@ krw@