udcf(4) degraded the sensor status from OK to WARNING on the first receiption error, which is to fast as this does not give ntpd the opportunity to use the sensor value.
If we decode a valid time information, it has to be available to applications
for some minimal time.  I choose a value of 5 minutes.  If we do not receive
any new valid time information during this period, the sensor will be degraded
to WARNING, and after another fifteen minutes it will be further degraded to
CRITICAL.

"makes sense to me" henning@, "sure" deraadt@