On frame buffers which endianness differs from the host CPU, we can not afford using ovbcopy() in the erasecols emulop, as the ovbcopy implementation might do larger-than-byte loads and store, which will not have the expected result if the source and destination addresses are not similarly aligned.
Instead, roll our own byte-only ovbcopy() in this case. This is made dependent
on a config(8) attribute to avoid bloating platforms which do not need this,
thus frame buffers which may set RI_BSWAP in ri_flg need to depend on this
attribute.

Problem spotted by matthieu@ using tcsh on sparc64 console.