First part of fixing broken beacon frames in acx(4) AP mode based on a diff from Sepherosa Ziehau (DragonFly);  The firmware TIM element template was initialized wrong.
Though, our ieee80211_alloc_beacon() routine calculates it's own TIM
element, which now results in having two TIMs in an acx(4) beacon.

Would it be an idea to introduce a ieee80211_alloc_beacon() flag to
turn the TIM element off?  Because fiddling out the TIM element in the
driver for cases in which the firmware calculates its own TIM is
a bit ugly ...

ok claudio@