After we bumped the maximal number of vnodes by quite a bit we became painfully aware of what the comment:    "This makes the algorithm O(n^2), but do you think I care?" actually means. I started to care.
Fix up the cache_purgevfs algorithm to not be O(n^2) since it's not
preemptible anyway and while I'm here, make this code actually return
the cache entries to the pool instead of hogging them and implement a
marginally faster free list. This way we return memory to the system
when some parameters change.

miod@ ok