Include the ioctl definitions for the modesetting interface.
This is so we can update libdrm, where some of the types in there depend
on these defines. This interface may well not be the one that ends up
being used when we finally have kernel modesetting working, so add a
comment to that effect.

after discussion with matthieu@ and miod@, ok miod@.