Add urndis(4), a driver for RNDIS Ethernet over USB.
It provides an Ethernet transport typically over EDGE or 3G on
cellphones similar to what cdce(4) does on other phones.  It is
likely to work with most of HTC's recent and coming Android based
phones but a bunch of other things (phones in particular) may use
it.

Started by me; brought to a working state by Jonathan Armani and
Fabien Romano over the past week, with some input and additions
from me.

Tested on i386 and macppc by me (HTC Hero), amd64 by Jonathan and
Fabien (HTC Hero), and by gilles@ (HTC Magic) on i386 or amd64.

It still has a few kinks to work out, but it works well enough that
I can commit this over my HTC Hero.

`commit it!' deraadt