Rewrite the main loop of the "sed s/..." command, to fix multiple issues regarding the replacement of zero-length strings.
This commit brings back rev. 1.16, but without the regression that
forced the backout:  No NUL bytes will be output now, not even when
the input file lacks a trailing newline character and there is a
zero-length match at the end.

OK otto@ deraadt@;
and naddy@ (who originally found the regression) checked that
the regression is indeed fixed.