Fix issues with AMD chipset ahci(4) attach functions. from Brad, ok jmatthew mikeb.
- Have the SB700 attach routine properly call the IDE to AHCI
function and migrate the ATI_SBX00_SATA_1 PCI id to using
the SB700 attach routine. The SB600 routine was being used
because it does the proper call. This just makes it so all
of the SB700 controllers use the same attach routine, especially
incase any other workarounds/errata type things come up or
not setting SB600 things for SB700.

- Have the Hudson-2 attach routine set the AHCI_F_IPMS_PROBE
flag to resolve the issue with Port Multiplier support. The
SB700 attachment function was being used because it was
setting the flag. This also fixes the other Hudson-2 AHCI
controllers so they won't experience the same issue.